refactor: extract jwt component
