+++ examples/access-control-migration/src/component/jwt-authentication-component.ts
@@ -1,0 +1,32 @@
+import {registerAuthenticationStrategy} from '@loopback/authentication';
+import {
+  Application,
+  Binding,
+  Component,
+  CoreBindings,
+  inject,
+} from '@loopback/core';
+import {JWTAuthenticationStrategy} from './jwt.auth.strategy';
+import {JWTService} from './jwt.service';
+import {
+  TokenServiceBindings,
+  TokenServiceConstants,
+  UserServiceBindings,
+} from './keys';
+import {MyUserService} from './user.service';
+
+export class JWTAuthenticationComponent implements Component {
+  bindings = [
+    Binding.bind(TokenServiceBindings.TOKEN_SECRET).to(
+      TokenServiceConstants.TOKEN_SECRET_VALUE,
+    ),
+    Binding.bind(TokenServiceBindings.TOKEN_EXPIRES_IN).to(
+      TokenServiceConstants.TOKEN_EXPIRES_IN_VALUE,
+    ),
+    Binding.bind(TokenServiceBindings.TOKEN_SERVICE).toClass(JWTService),
+    Binding.bind(UserServiceBindings.USER_SERVICE).toClass(MyUserService),
+  ];
+  constructor(@inject(CoreBindings.APPLICATION_INSTANCE) app: Application) {
+    registerAuthenticationStrategy(app, JWTAuthenticationStrategy);
+  }
+}
