+++ examples/access-control-migration/src/component/jwt.auth.strategy.ts
@@ -3,11 +3,11 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
+import {AuthenticationStrategy, TokenService} from '@loopback/authentication';
 import {inject} from '@loopback/context';
 import {HttpErrors, Request} from '@loopback/rest';
-import {AuthenticationStrategy, TokenService} from '@loopback/authentication';
 import {UserProfile} from '@loopback/security';
-import {TokenServiceBindings} from '../keys';
+import {TokenServiceBindings} from './keys';
 
 export class JWTAuthenticationStrategy implements AuthenticationStrategy {
   name = 'jwt';
