+++ examples/access-control-migration/src/component/jwt.service.ts
@@ -3,12 +3,12 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
+import {TokenService} from '@loopback/authentication';
 import {inject} from '@loopback/context';
 import {HttpErrors} from '@loopback/rest';
+import {securityId, UserProfile} from '@loopback/security';
 import {promisify} from 'util';
-import {TokenService} from '@loopback/authentication';
-import {UserProfile, securityId} from '@loopback/security';
-import {TokenServiceBindings} from '../keys';
+import {TokenServiceBindings} from './keys';
 
 const jwt = require('jsonwebtoken');
 const signAsync = promisify(jwt.sign);
