+++ examples/access-control-migration/src/component/user.service.ts
@@ -11,6 +11,10 @@ import {compare} from 'bcryptjs';
 import {User} from '../models/user.model';
 import {UserRepository} from '../repositories/user.repository';
 
+/**
+ * A pre-defined type for user credentials. It assumes a user logs in
+ * using the email and password. You can modify it if your app has different credential fields
+ */
 export type Credentials = {
   email: string;
   password: string;
