+++ examples/access-control-migration/src/keys.ts
@@ -3,32 +3,6 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
-import {TokenService, UserService} from '@loopback/authentication';
 import {BindingKey} from '@loopback/context';
-import {User} from './models';
-import {Credentials} from './services/user.service';
-
-export namespace TokenServiceConstants {
-  export const TOKEN_SECRET_VALUE = 'myjwts3cr3t';
-  export const TOKEN_EXPIRES_IN_VALUE = '21600';
-}
-
-export namespace TokenServiceBindings {
-  export const TOKEN_SECRET = BindingKey.create<string>(
-    'authentication.jwt.secret',
-  );
-  export const TOKEN_EXPIRES_IN = BindingKey.create<string>(
-    'authentication.jwt.expires.in.seconds',
-  );
-  export const TOKEN_SERVICE = BindingKey.create<TokenService>(
-    'services.authentication.jwt.tokenservice',
-  );
-}
-
-export namespace UserServiceBindings {
-  export const USER_SERVICE = BindingKey.create<UserService<User, Credentials>>(
-    'services.user.service',
-  );
-}
 
 export const RESOURCE_ID = BindingKey.create<string>('resourceId');
